requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.40.0
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.0
packaging==23.1
Pillow==10.0.0
pycairo==1.24.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.1
six==1.16.0
slackclient==2.9.4
soupsieve==2.4.1
urllib3==2.0.3
vectormath==0.2.2
Werkzeug==2.3.6
yarl==1.9.2
